ref: Move App code into its own folder
