+++ src/App/components/JsPsychExperiment.jsx
@@ -1,9 +1,9 @@
 import { initJsPsych } from "jspsych";
 import React, { useEffect, useMemo, useRef } from "react";
 
-import { config } from "../config/main";
-import { initParticipant } from "../firebase";
-import { buildTimeline, jsPsychOptions } from "../timelines/main";
+import { config } from "../../config/main";
+import { initParticipant } from "../../firebase";
+import { buildTimeline, jsPsychOptions } from "../../timelines/main";
 
 export default function JsPsychExperiment({
   studyId,
