+++ src/index.js
@@ -1,6 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App.jsx";
+
+import App from "./App/App.jsx";
+
+import "./App/index.css";
 
 ReactDOM.render(<App />, document.getElementById("root"));
